Move graph class into it's own library.
