+++ source/tit/sph/CMakeLists.txt
@@ -21,6 +21,7 @@ add_tit_library(
   DEPENDS
     tit::core
     tit::geom
+    tit::graph
 )
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
